+++ src/os/windows/alloc/heap.rs
@@ -5,21 +5,33 @@ use windows_sys::Win32::System::Memory::{
 };
 
 /// A handle to a process heap.
-pub type NSTDWindowsHeapHandle = NSTDInt;
+#[repr(C)]
+pub struct NSTDWindowsHeapHandle {
+    /// The private handle.
+    handle: NSTDInt,
+}
 
 /// Returns a handle to the default heap of the current process.
 ///
 /// # Returns
 ///
-/// `NSTDWindowsHeapHandle heap` - A handle to the default heap, null on error.
+/// `NSTDWindowsHeapHandle heap` - A handle to the default heap.
+///
+/// # Panics
+///
+/// Panics if getting a handle to the heap fails.
 ///
 /// # Safety
 ///
-/// See <https://docs.microsoft.com/en-us/windows/win32/api/heapapi/nf-heapapi-getprocessheap>.
+/// - See <https://docs.microsoft.com/en-us/windows/win32/api/heapapi/nf-heapapi-getprocessheap>.
+///
+/// - This operation can cause undefined behavior if it panics into non-Rust code.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_os_windows_alloc_heap_default() -> NSTDWindowsHeapHandle {
-    GetProcessHeap()
+    let handle = GetProcessHeap();
+    assert!(handle != 0);
+    NSTDWindowsHeapHandle { handle }
 }
 
 /// Creates a new private heap for the process.
@@ -33,13 +45,21 @@ pub unsafe extern "C" fn nstd_os_windows_alloc_heap_default() -> NSTDWindowsHeap
 ///
 /// `NSTDWindowsHeapHandle heap` - A handle to the new private heap.
 ///
+/// # Panics
+///
+/// Panics if creating a new heap fails.
+///
 /// # Safety
 ///
-/// See <https://docs.microsoft.com/en-us/windows/win32/api/heapapi/nf-heapapi-heapcreate>.
+/// - See <https://docs.microsoft.com/en-us/windows/win32/api/heapapi/nf-heapapi-heapcreate>.
+///
+/// - This operation can cause undefined behavior if it panics into non-Rust code.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_os_windows_alloc_heap_new(size: NSTDUInt) -> NSTDWindowsHeapHandle {
-    HeapCreate(0, size, 0)
+    let handle = HeapCreate(0, size, 0);
+    assert!(handle != 0);
+    NSTDWindowsHeapHandle { handle }
 }
 
 /// Allocates a block of memory on a heap.
@@ -63,7 +83,7 @@ pub unsafe extern "C" fn nstd_os_windows_alloc_heap_allocate(
     heap: NSTDWindowsHeapHandle,
     size: NSTDUInt,
 ) -> NSTDAnyMut {
-    HeapAlloc(heap, 0, size)
+    HeapAlloc(heap.handle, 0, size)
 }
 
 /// Allocates a zero-initialized block of memory on a heap.
@@ -87,7 +107,7 @@ pub unsafe extern "C" fn nstd_os_windows_alloc_heap_allocate_zeroed(
     heap: NSTDWindowsHeapHandle,
     size: NSTDUInt,
 ) -> NSTDAnyMut {
-    HeapAlloc(heap, HEAP_ZERO_MEMORY, size)
+    HeapAlloc(heap.handle, HEAP_ZERO_MEMORY, size)
 }
 
 /// Reallocates a block of memory on a heap.
@@ -114,7 +134,7 @@ pub unsafe extern "C" fn nstd_os_windows_alloc_heap_reallocate(
     ptr: &mut NSTDAnyMut,
     size: NSTDUInt,
 ) -> NSTDAllocError {
-    let new_mem = HeapReAlloc(heap, 0, *ptr, size);
+    let new_mem = HeapReAlloc(heap.handle, 0, *ptr, size);
     if !new_mem.is_null() {
         *ptr = new_mem;
         return NSTDAllocError::NSTD_ALLOC_ERROR_NONE;
@@ -143,7 +163,7 @@ pub unsafe extern "C" fn nstd_os_windows_alloc_heap_deallocate(
     heap: NSTDWindowsHeapHandle,
     ptr: &mut NSTDAnyMut,
 ) -> NSTDAllocError {
-    if HeapFree(heap, 0, *ptr) != 0 {
+    if HeapFree(heap.handle, 0, *ptr) != 0 {
         *ptr = NSTD_NULL;
         return NSTDAllocError::NSTD_ALLOC_ERROR_NONE;
     }
@@ -154,7 +174,7 @@ pub unsafe extern "C" fn nstd_os_windows_alloc_heap_deallocate(
 ///
 /// # Parameters:
 ///
-/// - `NSTDWindowsHeapHandle *heap` - The heap to destroy.
+/// - `NSTDWindowsHeapHandle heap` - The heap to destroy.
 ///
 /// # Returns
 ///
@@ -166,10 +186,9 @@ pub unsafe extern "C" fn nstd_os_windows_alloc_heap_deallocate(
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_os_windows_alloc_heap_free(
-    heap: &mut NSTDWindowsHeapHandle,
+    heap: NSTDWindowsHeapHandle,
 ) -> NSTDAllocError {
-    if HeapDestroy(*heap) != 0 {
-        *heap = 0;
+    if HeapDestroy(heap.handle) != 0 {
         return NSTDAllocError::NSTD_ALLOC_ERROR_NONE;
     }
     NSTDAllocError::NSTD_ALLOC_ERROR_HEAP_NOT_FOUND